src/main/java/ar/edu/unlam/tallerweb1/servicios/ServicioUsuario.java
package ar.edu.unlam.tallerweb1.servicios;

import ar.edu.unlam.tallerweb1.controladores.DatosRegistro;
import ar.edu.unlam.tallerweb1.modelo.Usuario;
import ar.edu.unlam.tallerweb1.repositorios.TablaUsuario;

public class ServicioUsuario {

    private TablaUsuario tablaUsuario = TablaUsuario.getInstance();

    public Usuario registrar(DatosRegistro datosRegistro) {
        if(lasClavesSonDistintas(datosRegistro)){
            throw new ClavesDistintasException();
        }
        if(laClaveTieneLongitudIncorrecta(datosRegistro)){
            throw new ClaveLongitudIncorrectaException();
        }
        if(tablaUsuario.existeUsuarioCon(datosRegistro.getEmail())){
            throw new UsuarioYaExisteException();
        }
        Usuario nuevoUsuario = new Usuario(datosRegistro);
        tablaUsuario.agregar(nuevoUsuario);
        return nuevoUsuario;
    }

    private boolean lasClavesSonDistintas(DatosRegistro datosRegistro) {
        return !datosRegistro.getClave().equals(datosRegistro.getRepiteClave());
    }

    private boolean laClaveTieneLongitudIncorrecta(DatosRegistro datosRegistro) {
        return datosRegistro.getClave().length() < 8;
    }
}
